pyproject.toml
# ============================================================
# AI Novel Generator 4.5 - 项目配置
# 版权所有 © 2026 新疆幻城网安科技有限责任公司 (幻城科技)
# ============================================================

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-novel-generator"
version = "4.5.0"
description = "AI Novel Generator 4.5 - 智能连贯性系统 + 22+提供商支持"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "幻城科技", email = "contact@example.com"}
]
keywords = ["ai", "novel", "writing", "llm", "gradio", "gpt"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Operating System :: OS Independent",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Utilities",
]

dependencies = [
    # 核心依赖
    "gradio>=4.44.0,<5.0.0",
    "openai>=1.54.0,<2.0.0",
    "python-docx>=1.1.2,<2.0.0",

    # 间接依赖（显式声明）
    "httpx>=0.24.0,<0.28.0",
    "httpcore>=0.18.0,<1.1.0",
    "aiohttp>=3.8.0,<3.12.0",
    "huggingface_hub>=0.19.0,<0.25.0",
    "gradio_client>=0.16.0,<1.1.0",
    "pandas>=1.5.0,<3.0.0",
    "numpy>=1.23.0,<2.2.0",
    "pillow>=9.5.0,<12.0.0",
    "python-multipart>=0.0.5,<0.0.20",
    "jinja2>=3.1.0,<4.0.0",
    "markupsafe>=2.1.0,<3.0.0",
    "typing-extensions>=4.5.0,<5.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-core>=2.0.0,<3.0.0",
    "fastapi>=0.100.0,<0.120.0",
    "uvicorn[standard]>=0.23.0,<0.33.0",
    "websockets>=11.0,<14.0",
    "python-dateutil>=2.8.0,<3.0.0",
    "tqdm>=4.65.0,<5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-cov>=4.1.0,<6.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "ipython>=8.12.0",
    "ipdb>=0.13.0",
    "pip-audit>=2.5.0",
]

all = ["ai-novel-generator[dev]"]

[project.urls]
Homepage = "https://github.com/example/ai-novel-generator"
Documentation = "https://github.com/example/ai-novel-generator/docs"
Repository = "https://github.com/example/ai-novel-generator"
"Bug Tracker" = "https://github.com/example/ai-novel-generator/issues"

[project.scripts]
ai-novel-generator = "run:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["src"]
include = ["src*"]

# ------------------- 测试配置 -------------------

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# ------------------- Ruff 配置 -------------------

[tool.ruff]
target-version = "py39"
line-length = 100
indent-width = 4

# 包含的文件
include = ["*.py", "*.pyi", "**/pyproject.toml"]

# 排除的目录
exclude = [
    ".git",
    ".venv",
    "venv",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
# 启用的检查器
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]

# 忽略的检查器
ignore = [
    "E501",   # line too long (由 formatter 处理)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "SIM108", # Use ternary operator (可读性考虑)
]

# 每行允许的未使用导入
unfixable = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-single-line = false
lines-after-imports = 2

# ------------------- MyPy 配置 -------------------

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true
show_error_context = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = "gradio.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "openai.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "docx.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "gradio_client.*"
ignore_missing_imports = true

# ------------------- Coverage 配置 -------------------

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
